mcleaned up do client
