--- src/clients/digital_ocean/models/droplet_create.rs
@@ -1,14 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-use super::image::Image;
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct DropletCreate {
-    name: String,
-    region: String,
-    size: String,
-    image: Image,
-    tags: Vec<String>,
-    ssh_key: Option<Vec<String>>,
-    backup: Option<bool>,
-}
