+++ src/clients/digital_ocean/models/mod.rs
@@ -1,4 +1,4 @@
-pub mod droplet_create;
+pub mod droplet;
 pub mod image;
 pub mod networks;
 pub mod region;
